Reorganized screens
All screen components no longer have the 'screen' suffix
All components have their own folder with a separate styles.ts file
Updated status bar to show text in dark mode
