+++ MediWay/Main.tsx
@@ -1,17 +1,21 @@
 import React, { useContext } from 'react';
 import AuthStack from './navigation/AuthStack';
-import HomeStack from './navigation/HomeStack';
+import TabsStack from './navigation/TabsStack';
 import GetStartedStack from './navigation/GetStartedStack';
 import { AuthContext } from './contexts/AuthContext';
+import { StatusBar } from 'react-native';
 
 const Main = () => {
     const { state: { loggedIn, ready } } = useContext(AuthContext);
 
     return (
         <>
+            <StatusBar
+                barStyle="dark-content"
+            />
             {ready ? (
                 loggedIn ? (
-                    <HomeStack />
+                    <TabsStack />
                 ) : (
                     <AuthStack />
                 )
