+++ MediWay/assets/interfaces.ts
@@ -1,20 +1,14 @@
-export interface MessageProp {
-    id: string
-    isIncoming: boolean,
-    text: string
-}
-
 export interface Profile {
-  email: string;
+    email: string;
 }
 
 export interface User {
-  username: string;
-  profile: Profile;
+    username: string;
+    profile: Profile;
 }
 
 export interface AuthState {
-  ready: boolean;
-  loggedIn: boolean;
-  initialRoute: string;
+    ready: boolean;
+    loggedIn: boolean;
+    initialRoute: string;
 }
