+++ MediWay/navigation/GetStartedStack.tsx
@@ -1,7 +1,7 @@
 //used only for displaying and navigatiion for get started screens
 import { createStackNavigator } from '@react-navigation/stack';
 import GetStarted from '../screens/GetStarted/GetStarted';
-import SplashScreen from '../screens/SplashScreen/SplashScreen';
+import Splash from '../screens/Splash/Splash';
 import MoreInfo from '../screens/MoreInfo/MoreInfo';
 
 const Stack = createStackNavigator();
@@ -13,7 +13,7 @@ function GetStartedStack() {
                 headerShown: false, // This removes the header completely
             }}
         >
-            <Stack.Screen name="SplashScreen" component={SplashScreen} />
+            <Stack.Screen name="SplashScreen" component={Splash} />
             <Stack.Screen name="GetStarted" component={GetStarted} />
             <Stack.Screen name="MoreInfo" component={MoreInfo} />
         </Stack.Navigator>
