+++ MediWay/screens/ForgotPassword/ForgotPassword.tsx
@@ -11,9 +11,9 @@ import {
   ScrollView,
   Image,
 } from 'react-native';
-import Button from '../components/Button/Button';
-import { COLORS } from '../assets/constants';
-import { StyleSheet } from 'react-native';
+import Button from '../../components/Button/Button';
+import { ENDPOINTS } from '../../assets/api';
+import styles from './styles';
 
 interface ForgotPasswordProps {
   onBack?: () => void;
@@ -22,7 +22,6 @@ interface ForgotPasswordProps {
 }
 
 const ForgotPassword: React.FC<ForgotPasswordProps> = ({
-  onBack,
   onLoginPress,
   onCodeSent,
 }) => {
@@ -37,7 +36,7 @@ const ForgotPassword: React.FC<ForgotPasswordProps> = ({
 
     setLoading(true);
     try {
-      const response = await fetch('http://localhost:8000/api/v1/auth/forgot-password', {
+      const response = await fetch(ENDPOINTS.forgotPassword, {
         method: 'POST',
         headers: {
           'Content-Type': 'application/json',
@@ -79,7 +78,7 @@ const ForgotPassword: React.FC<ForgotPasswordProps> = ({
         <ScrollView contentContainerStyle={styles.scrollContent}>
           <View style={styles.header}>
             <Image
-              source={require('../assets/images/logo.png')}
+              source={require('../../assets/images/logo.png')}
               style={styles.logo}
               resizeMode="contain"
             />
@@ -123,77 +122,4 @@ const ForgotPassword: React.FC<ForgotPasswordProps> = ({
   );
 };
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: COLORS.WHITE,
-  },
-  keyboardView: {
-    flex: 1,
-  },
-  scrollContent: {
-    flexGrow: 1,
-    justifyContent: 'center',
-    padding: 20,
-  },
-  header: {
-    alignItems: 'center',
-    marginBottom: 40,
-  },
-  logo: {
-    width: 200,
-    height: 200,
-    marginBottom: -30,
-  },
-  title: {
-    fontSize: 28,
-    fontWeight: 'bold',
-    color: COLORS.BLACK,
-    marginBottom: 16,
-  },
-  subtitle: {
-    fontSize: 16,
-    color: '#00000080',
-    textAlign: 'center',
-    lineHeight: 24,
-  },
-  form: {
-    width: '100%',
-  },
-  inputContainer: {
-    marginBottom: 30,
-  },
-  label: {
-    fontSize: 16,
-    fontWeight: '500',
-    color: COLORS.BLACK,
-    marginBottom: 8,
-  },
-  input: {
-    borderWidth: 1,
-    borderColor: '#00000030',
-    borderRadius: 12,
-    padding: 16,
-    fontSize: 16,
-    backgroundColor: COLORS.WHITE,
-  },
-  disabledButton: {
-    opacity: 0.6,
-  },
-  loginContainer: {
-    flexDirection: 'row',
-    justifyContent: 'center',
-    marginTop: 20,
-  },
-  loginText: {
-    fontSize: 14,
-    color: '#00000080',
-  },
-  loginLink: {
-    fontSize: 14,
-    color: COLORS.TERTIARY,
-    fontWeight: '500',
-  },
-});
-
-export default ForgotPassword; 
+export default ForgotPassword;
