+++ MediWay/screens/Register/Register.tsx
@@ -11,17 +11,16 @@ import {
     ScrollView,
     Image,
 } from 'react-native';
-import Button from '../components/Button/Button';
-import { COLORS } from '../assets/constants';
-import { StyleSheet } from 'react-native';
-import { AuthContext } from '../contexts/AuthContext';
+import Button from '../../components/Button/Button';
+import { AuthContext } from '../../contexts/AuthContext';
+import styles from './styles';
 
 interface RegisterProps {
     onBack?: () => void;
     onLoginPress?: () => void;
 }
 
-const Register: React.FC<RegisterProps> = ({
+const RegisterScreen: React.FC<RegisterProps> = ({
     onLoginPress,
 }) => {
     const [formData, setFormData] = useState({
@@ -125,7 +124,7 @@ const Register: React.FC<RegisterProps> = ({
                 <ScrollView contentContainerStyle={styles.scrollContent}>
                     <View style={styles.header}>
                         <Image
-                            source={require('../assets/images/logo.png')}
+                            source={require('../../assets/images/logo.png')}
                             style={styles.logo}
                             resizeMode="contain"
                         />
@@ -239,75 +238,4 @@ const Register: React.FC<RegisterProps> = ({
     );
 };
 
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        backgroundColor: COLORS.WHITE,
-    },
-    keyboardView: {
-        flex: 1,
-    },
-    scrollContent: {
-        flexGrow: 1,
-        padding: 20,
-        paddingTop: 40,
-    },
-    header: {
-        alignItems: 'center',
-        marginBottom: 30,
-    },
-    logo: {
-        width: 200,
-        height: 200,
-        marginBottom: -30,
-    },
-    title: {
-        fontSize: 28,
-        fontWeight: 'bold',
-        color: COLORS.BLACK,
-        marginBottom: 8,
-    },
-    subtitle: {
-        fontSize: 16,
-        color: '#00000080',
-    },
-    form: {
-        width: '100%',
-    },
-    inputContainer: {
-        marginBottom: 16,
-    },
-    label: {
-        fontSize: 16,
-        fontWeight: '500',
-        color: COLORS.BLACK,
-        marginBottom: 8,
-    },
-    input: {
-        borderWidth: 1,
-        borderColor: '#00000030',
-        borderRadius: 12,
-        padding: 16,
-        fontSize: 16,
-        backgroundColor: COLORS.WHITE,
-    },
-    disabledButton: {
-        opacity: 0.6,
-    },
-    loginContainer: {
-        flexDirection: 'row',
-        justifyContent: 'center',
-        marginTop: 20,
-    },
-    loginText: {
-        fontSize: 14,
-        color: '#00000080',
-    },
-    loginLink: {
-        fontSize: 14,
-        color: COLORS.TERTIARY,
-        fontWeight: '500',
-    },
-});
-
-export default Register;
+export default RegisterScreen;
