--- MediWay/screens/Test1.tsx
@@ -1,9 +1,0 @@
-import { Text } from "react-native-gesture-handler";
-
-const Test1 = () => {
-    return (
-        <Text>Hello Test1</Text>
-    );
-};
-
-export default Test1;
